requirements.txt
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
pastel==0.2.1
Pillow==9.2.0
pylev==1.4.0
